+++ data/src/main/java/com/moneymong/moneymong/data/di/DataSourceModule.kt
@@ -2,6 +2,8 @@ package com.moneymong.moneymong.data.di
 
 import com.moneymong.moneymong.data.datasource.ledger.LedgerRemoteDataSource
 import com.moneymong.moneymong.data.datasource.ledger.LedgerRemoteDataSourceImpl
+import com.moneymong.moneymong.data.datasource.ledgerdetail.LedgerDetailRemoteDataSource
+import com.moneymong.moneymong.data.datasource.ledgerdetail.LedgerDetailRemoteDataSourceImpl
 import com.moneymong.moneymong.data.datasource.ocr.OCRRemoteDataSource
 import com.moneymong.moneymong.data.datasource.ocr.OCRRemoteDataSourceImpl
 import dagger.Binds
@@ -22,4 +24,9 @@ interface DataSourceModule {
     fun provideLedgerRemoteDataSource(
         ledgerRemoteDataSourceImpl: LedgerRemoteDataSourceImpl
     ): LedgerRemoteDataSource
+
+    @Binds
+    fun provideLedgerDetailRemoteDataSource(
+        ledgerDetailRemoteDataSourceImpl: LedgerDetailRemoteDataSourceImpl
+    ): LedgerDetailRemoteDataSource
 }
