+++ data/src/main/java/com/moneymong/moneymong/data/repository/ledger/LedgerRepositoryImpl.kt
@@ -3,17 +3,10 @@ package com.moneymong.moneymong.data.repository.ledger
 import com.moneymong.moneymong.data.datasource.ledger.LedgerRemoteDataSource
 import com.moneymong.moneymong.data.mapper.ledger.toEntity
 import com.moneymong.moneymong.data.mapper.ledger.toRequest
-import com.moneymong.moneymong.data.mapper.ledgerdetail.toRequest
-import com.moneymong.moneymong.domain.entity.ledger.LedgerTransactionDetailEntity
 import com.moneymong.moneymong.domain.entity.ledger.LedgerTransactionEntity
 import com.moneymong.moneymong.domain.entity.ledger.LedgerTransactionListEntity
 import com.moneymong.moneymong.domain.param.ledger.LedgerTransactionListParam
 import com.moneymong.moneymong.domain.param.ledger.LedgerTransactionParam
-import com.moneymong.moneymong.domain.param.ledgerdetail.DeleteLedgerDocumentParam
-import com.moneymong.moneymong.domain.param.ledgerdetail.DeleteLedgerReceiptParam
-import com.moneymong.moneymong.domain.param.ledgerdetail.LedgerDocumentParam
-import com.moneymong.moneymong.domain.param.ledgerdetail.LedgerReceiptParam
-import com.moneymong.moneymong.domain.param.ledgerdetail.LedgerTransactionDetailParam
 import com.moneymong.moneymong.domain.repository.ledger.LedgerRepository
 import javax.inject.Inject
 
@@ -35,32 +28,4 @@ class LedgerRepositoryImpl @Inject constructor(
     override suspend fun postLedgerTransaction(param: LedgerTransactionParam): Result<LedgerTransactionEntity> =
         ledgerRemoteDataSource.postLedgerTransaction(id = param.id, body = param.toRequest())
             .map { it.toEntity() }
-
-    override suspend fun fetchLedgerTransactionDetail(detailId: Int): Result<LedgerTransactionDetailEntity> =
-        ledgerRemoteDataSource.fetchLedgerTransactionDetail(detailId = detailId)
-            .map { it.toEntity() }
-
-    override suspend fun postLedgerReceiptTransaction(body: LedgerReceiptParam): Result<Unit> =
-        ledgerRemoteDataSource.postLedgerReceiptTransaction(detailId = body.detailId, body = body.toRequest())
-
-    override suspend fun postLedgerDocumentTransaction(body: LedgerDocumentParam): Result<Unit> =
-        ledgerRemoteDataSource.postLedgerDocumentTransaction(detailId = body.detailId, body = body.toRequest())
-
-    override suspend fun updateLedgerTransactionDetail(body: LedgerTransactionDetailParam): Result<LedgerTransactionDetailEntity> =
-        ledgerRemoteDataSource.updateLedgerTransactionDetail(
-            detailId = body.detailId,
-            body = body.toRequest()
-        ).map { it.toEntity() }
-
-    override suspend fun deleteLedgerReceiptTransaction(deleteLedgerReceiptParam: DeleteLedgerReceiptParam): Result<Unit> =
-        ledgerRemoteDataSource.deleteLedgerReceiptTransaction(
-            detailId = deleteLedgerReceiptParam.detailId,
-            receiptId = deleteLedgerReceiptParam.receiptId
-        )
-
-    override suspend fun deleteLedgerDocumentTransaction(deleteLedgerDocumentParam: DeleteLedgerDocumentParam): Result<Unit> =
-        ledgerRemoteDataSource.deleteLedgerDocumentTransaction(
-            detailId = deleteLedgerDocumentParam.detailId,
-            documentId = deleteLedgerDocumentParam.documentId
-        )
 }
